[BPK-1920] Update dependencies and minSDK

+++ Backpack/src/main/java/net/skyscanner/backpack/text/BpkText.kt
@@ -14,7 +14,7 @@ open class BpkText(
   attrs: AttributeSet?,
   defStyleAttr: Int) : AppCompatTextView(context, attrs, defStyleAttr) {
 
-  @IntDef(XS.toLong(), SM.toLong(), BASE.toLong(), LG.toLong(), XL.toLong(), XXL.toLong())
+  @IntDef(XS, SM, BASE, LG, XL, XXL)
 
   annotation class Styles
